fixing bulk push to redis

+++ app/workers/reprocess_paper.rb
@@ -3,6 +3,7 @@ class ReprocessPaper
 
   def perform(doi)
     paper = PaperResult.find_by(doi: doi)
+    return if paper.nil?
     paper.destroy!
     AnalyzePaper.perform_async(doi)
   end
